\name{plot.predict.GORMC}
\alias{plot.predict.GORMC}
\title{
Plot the predicted survival curves.
}
\description{
Plot the predicted survival curve of the new subject.
}
\usage{
\method{plot}{predict.GORMC}(x, ...)
}
\arguments{
  \item{x}{
An object from the predict.GORMC function.
}
  \item{\dots}{
Other plot arguments.
}
}

