\name{summary.GORMC}
\alias{summary.GORMC}
\title{
Summary table of hypothesis tests for the coefficients in the GORMC model.
}
\description{
Estimates, standard errors, test statistics and p values are presented for each coefficient in the GORMC model. The ``INC"
 indicate the covariate is in the cure rate (incidence) part, ``LAT" indicate the covariate is in the survival (latency)
 part.
}
\usage{
\method{summary}{GORMC}(object, ...)
}
\arguments{
  \item{object}{
An object from the function GORMC.
}
  \item{\dots}{
Other auguments to be specified.
}
}

